agents/networking/pyproject.toml
[project]
name = "networking"
version = "0.1.0"
description = "networking-copilot using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.177.0,<1.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "httpx>=0.27.0,<1.0.0",
    "pyyaml>=6.0.0,<7.0.0"
]

[project.scripts]
networking = "networking.main:run"
run_crew = "networking.main:run"
train = "networking.main:train"
replay = "networking.main:replay"
test = "networking.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
